+++ oneflow/python/framework/graph_build_util.py
@@ -23,10 +23,11 @@ import oneflow.core.job.scope_pb2 as scope_pb2_util
 import oneflow.python.framework.attr_util as attr_util
 import oneflow.python.framework.c_api_util as c_api_util
 import oneflow.python.framework.placement_util as placement_util
-import oneflow.python.framework.runtime_mode as runtime_mode
 import oneflow.python.framework.scope_util as scope_util
 import oneflow.python.framework.session_context as session_context
 import oneflow._oneflow_internal
+from oneflow._oneflow_internal import Tensor as InternalTensor
+from oneflow.python.framework.tensor import Tensor
 
 
 lazy_mode = oneflow._oneflow_internal.lazy_mode
@@ -125,3 +126,61 @@ def make_new_block_scope(prev_scope, block):
 
 def scope_to_proto(scope):
     return text_format.Parse(scope._proto_str, scope_pb2_util.ScopeProto())
+
+
+def build_graph_input_arg(op_name, arg):
+    assert isinstance(arg, (Tensor, InternalTensor))
+    input_conf = (
+        oneflow._oneflow_internal.oneflow.core.operator.op_conf.FeedInputOpConf()
+    )
+
+    input_op = oneflow._oneflow_internal.one.FeedInputOpExpr(
+        op_name, input_conf, ["in_0"], ["out_0"]
+    )
+    attrs = oneflow._oneflow_internal.MutableCfgAttrMap()
+
+    if isinstance(arg, Tensor):
+        if not arg.is_determined:
+            arg.determine()
+        tensor_in_c = arg._local_or_consistent_tensor
+    else:
+        tensor_in_c = arg
+
+    lazy_arg = input_op.apply([tensor_in_c], attrs)[0]
+    return lazy_arg
+
+
+def build_graph_state(op_name, state_tensor):
+    var_conf = (
+        oneflow._oneflow_internal.oneflow.core.operator.op_conf.FeedVariableOpConf()
+    )
+
+    var_op = oneflow._oneflow_internal.one.FeedVariableOpExpr(
+        op_name, var_conf, ["in_0"], ["out_0"]
+    )
+    attrs = oneflow._oneflow_internal.MutableCfgAttrMap()
+
+    assert isinstance(state_tensor, Tensor)
+    if not state_tensor.is_determined:
+        state_tensor.determine()
+    tensor_in_c = state_tensor._local_or_consistent_tensor
+
+    lazy_tensor = var_op.apply([tensor_in_c], attrs)[0]
+    return lazy_tensor
+
+
+def build_graph_output(op_name, out):
+    assert isinstance(out, InternalTensor)
+    assert out.is_lazy
+
+    output_conf = (
+        oneflow._oneflow_internal.oneflow.core.operator.op_conf.FetchOutputOpConf()
+    )
+
+    output_op = oneflow._oneflow_internal.one.FetchOutputOpExpr(
+        op_name, output_conf, ["in_0"], ["out_0"]
+    )
+    attrs = oneflow._oneflow_internal.MutableCfgAttrMap()
+
+    eager_out = output_op.apply([out], attrs)[0]
+    return eager_out
