+++ oneflow/python/test/graph/test_user_op_expr.py
@@ -18,12 +18,6 @@ import unittest
 import numpy as np
 import os
 
-os.environ["MASTER_ADDR"] = "127.0.0.1"
-os.environ["MASTER_PORT"] = "12139"
-os.environ["WORLD_SIZE"] = "1"
-os.environ["RANK"] = "0"
-os.environ["LOCAL_RANK"] = "0"
-
 import oneflow
 import oneflow.experimental as flow
 import oneflow.python.framework.session_context as session_ctx
